+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- AnalyticsManager module
+- PrivacyManager module
+
 ## [0.1.2] - 2022-01-26
 
 ### Fixed
@@ -13,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.1] - 2022-01-23
 
-This is purely an administrative release -- no code changes
+- This is purely an administrative release -- no code changes
 
 ## [0.1.0] - 2022-01-23